Factorize methods related to local repo
+ Implement a watch on local repo to cache artifacts and refresh when
repo is modified
