+++ org.eclipse.lsp4xml.extensions.maven/src/main/java/org/eclipse/lsp4xml/extensions/maven/searcher/ArtifactSearcherManager.java
@@ -17,7 +17,7 @@ public class ArtifactSearcherManager {
 	private final IArtifactSearcher localSearcher;
 
 	public ArtifactSearcherManager() {
-		localSearcher = new LocalArtifactSearcher();
+		localSearcher = new LocalRepositorySearcher();
 	}
 
 	public static ArtifactSearcherManager getInstance() {
